Adding support for language-specific inflection table indices. Now when an Inflection Table is requested, the index will display if a word isn't selected, or if the part of speech of the selected word does not have a corresponding inflection table.  A link back to the Inflection Table index will also appear on every inflection table display.
The inflection table index is expected to adhere to TEI P5, formatted as div of type 'contents' in the front  matter. (/TEI/text/front/div[@type='contents'])

The syntax for the ptr elements for each item in the list is provisional until we have established our linking architecture. For now, it's expected to adhere to "#<pofs>[|<param_name:param_value]+]